pyproject.toml
[tool.poetry]
name = "inventory-management-api"
version = "0.1.0"
description = ""
authors = ["Hailemichael Atrsaw Tibebu <hailat.alx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
